requirements.txt
numpy>=1.21.0
pandas>=1.3.0
PyYAML>=6.0
tqdm>=4.62.0
SALib>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0

